requirements.txt
setuptools>=65.5.1
wheel==0.43.0
psycopg2==2.9.9
sqlalchemy>=1.4.52
sqlalchemy-redshift>=0.8.14
python-dotenv>=1.0.1
pandas==2.1.4
sqlfluff==3.1.0
alembic==1.13.2

boto3~=1.34.154
kimballorm~=0.1.0
pytest~=8.3.2

pyodbc~=5.1.0
